+++ client/web/dev/server/development.server.ts
@@ -1,5 +1,4 @@
 import chalk from 'chalk'
-import { Application } from 'express'
 import { once } from 'lodash'
 import signale from 'signale'
 import createWebpackCompiler, { Configuration } from 'webpack'
@@ -34,11 +33,10 @@ export async function startDevelopmentServer(): Promise<void> {
         }),
     }
 
-    // It's not possible to use `WebpackDevServer.Configuration` here yet, because
-    // type definitions for the `webpack-dev-server` are not updated to match v4.
-    const developmentServerConfig = {
+    const developmentServerConfig: WebpackDevServer.Configuration = {
         // react-refresh plugin triggers page reload if needed.
         liveReload: false,
+        allowedHosts: 'all',
         hot: IS_HOT_RELOAD_ENABLED,
         // TODO: resolve https://github.com/webpack/webpack-dev-server/issues/2313 and enable HTTPS.
         https: false,
@@ -53,15 +51,14 @@ export async function startDevelopmentServer(): Promise<void> {
             directory: STATIC_ASSETS_PATH,
             publicPath: [STATIC_ASSETS_URL, '/'],
         },
-        firewall: false,
         proxy: [proxyConfig],
-        onBeforeSetupMiddleware(app: Application) {
-            app.use(getCSRFTokenCookieMiddleware(csrfCookieValue))
+        onBeforeSetupMiddleware: developmentServer => {
+            developmentServer.app.use(getCSRFTokenCookieMiddleware(csrfCookieValue))
         },
     }
 
     const compiler = createWebpackCompiler(webpackConfig)
-    const server = new WebpackDevServer(compiler, developmentServerConfig as WebpackDevServer.Configuration)
+    const server = new WebpackDevServer(developmentServerConfig, compiler)
 
     compiler.hooks.afterEmit.tap(
         'development-server-logger',